chore(agent,pkg): update mode's name when running for the host machine
Currently, Agent has two operational modes: host, which runs targeting
the currenty OS running it, compiled or insider a Docker Container, and
connector, which accesses the containers on the machine, gets the
information needed, and turns it into an in-memory agent.

However, "host" mode isn't a good name because we already use this term
to mean that the agent was compiled to run directly on the machine,
which could cause misunderstandings about this mode.

To solve this problem, we have renamed the "host" mode to "classic",
letting the term "host" only differentiate between Agent's running
inside Docker Container or not.

Classic mode means that the Agent's SSH server runs for the host
machine, being compiled or embedded inside a Docker Container, using the
host "/etc/passwd", "/etc/shadow", redirecting the SSH's connection to
the devices STDIN, STDOUT and STDERR, and anything else necessary.
