+++ pkg/agent/server/modes/classic/command/command.go
