+++ pkg/agent/server/modes/classic/pty.go
@@ -1,4 +1,4 @@
-package host
+package classic
 
 import (
 	"io"
